adding files

+++ include/Initializer.h
@@ -45,15 +45,15 @@ public:
 
 private:
 
-    void FindHomography(vector<bool> &vbMatchesInliers, float &score, cv::Mat &H21);
-    void FindFundamental(vector<bool> &vbInliers, float &score, cv::Mat &F21);
+    void FindHomography(vector<bool> &vbMatchesInliers, double &score, cv::Mat &H21);
+    void FindFundamental(vector<bool> &vbInliers, double &score, cv::Mat &F21);
 
     cv::Mat ComputeH21(const vector<cv::Point2f> &vP1, const vector<cv::Point2f> &vP2);
     cv::Mat ComputeF21(const vector<cv::Point2f> &vP1, const vector<cv::Point2f> &vP2);
 
-    float CheckHomography(const cv::Mat &H21, const cv::Mat &H12, vector<bool> &vbMatchesInliers, float sigma);
+    double CheckHomography(const cv::Mat &H21, const cv::Mat &H12, vector<bool> &vbMatchesInliers, double sigma);
 
-    float CheckFundamental(const cv::Mat &F21, vector<bool> &vbMatchesInliers, float sigma);
+    double CheckFundamental(const cv::Mat &F21, vector<bool> &vbMatchesInliers, double sigma);
 
     bool ReconstructF(vector<bool> &vbMatchesInliers, cv::Mat &F21, cv::Mat &K,
                       cv::Mat &R21, cv::Mat &t21, vector<cv::Point3f> &vP3D, vector<bool> &vbTriangulated, float minParallax, int minTriangulated);
@@ -67,7 +67,7 @@ private:
 
     int CheckRT(const cv::Mat &R, const cv::Mat &t, const vector<cv::KeyPoint> &vKeys1, const vector<cv::KeyPoint> &vKeys2,
                        const vector<Match> &vMatches12, vector<bool> &vbInliers,
-                       const cv::Mat &K, vector<cv::Point3f> &vP3D, float th2, vector<bool> &vbGood, float &parallax);
+                       const cv::Mat &K, vector<cv::Point3f> &vP3D, double th2, vector<bool> &vbGood, double &parallax);
 
     void DecomposeE(const cv::Mat &E, cv::Mat &R1, cv::Mat &R2, cv::Mat &t);
 
@@ -86,7 +86,7 @@ private:
     cv::Mat mK;
 
     // Standard Deviation and Variance
-    float mSigma, mSigma2;
+    double mSigma, mSigma2;
 
     // Ransac max iterations
     int mMaxIterations;
